🔀 Dev/v0.10.3 (#35) Feature Update
* .help by default lists all the plugins with commands you can use

Added a second level of filtering to .help so you can filter commands with a plugin or second filter for the command search
.help now has commands and plugins sorted
Adds .help-all which functions similar to previous .help except its ordered
Added new unit tests

* Commands can now be overloaded.

If there is ambiguity in which command was intended to be called then you will be presented with the options and asked to pick one with .#
Can specify a command for a specific plugin by using the name of it first before the command

* .help by default lists all the plugins with commands you can use

Added a second level of filtering to .help so you can filter commands with a plugin or second filter for the command search
.help now has commands and plugins sorted
Adds .help-all which functions similar to previous .help except its ordered
Added new unit tests

* When you submit a command but its not quite right up to three of the closest possible command matches are returned.

Updating presentation of messages and help.
Updating unit tests for the new presentation of messages

* Command History and Recall Feature (#33)

* When you submit a command but its not quite right up to three of the closest possible command matches are returned.

Updating presentation of messages and help.
Updating unit tests for the new presentation of messages

* Adds the ability to repeat any of the last 10 commands.  Can use .! to repeat the previous one, .! [#] to repeat the # back command, and .! list to list out the last 10 command history for the user.

* Unit tests added for repeat commands

* Updating README.md reflecting all changes

Added additional colors correlating to various colors in VRising

* Adding reference to V Roles in the Middleware section as an example plugin already using the middleware.

* Guess different version of .net is being used on the GitHub Action workflow than in the project