🔀 Fix command resolution when assembly name matches command prefix (#36)
An assembly name which matches the start of a command name before a space would block out those commands if you didn't specify the assembly name they were in.  Now if the command being looked for isn't found in the assembly go back to look across all the commands without specifying an assembly.  In future, potentially should handle the commands being overloaded in this case but this is a fairly rare case only brought up due to killfeed prefixing its commands with killfeed. (#36)